+++ src/main/java/com/calorator/repository/UserRepository.java
@@ -9,11 +9,11 @@ import java.util.Optional;
 @Repository
 public interface UserRepository {
     void save(UserEntity userEntity);
-    UserEntity findById(Long id);
+    UserEntity findById(int id);
     UserEntity findByName(String name);
     void update(UserEntity userEntity);
     List<UserEntity> findAll();
-    void delete(Long id);
+    void delete(int id);
     Optional<UserEntity> findByEmail(String email);
     Optional<UserEntity> findByUserName(String username);
     Optional<UserEntity> findByPassword(String password);
